Move CRD client to the k8s package
